Move EncodingNotFoundException to src/Smalot/PdfParser/Exception
